Feature/day 18

+++ app/Http/Controllers/ProjectController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Project\ProjectStoreRequest;
 use App\Http\Requests\Project\ProjectUpdateRequest;
 use App\Models\Project;
+use App\Models\User;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
 class ProjectController extends Controller
@@ -17,7 +19,9 @@ class ProjectController extends Controller
      */
     public function index(): View
     {
-        $projects = Project::all();
+        Gate::authorize('viewAny', Project::class);
+
+        $projects = Project::orderBy('created_at')->get();
 
         return view('pages.projects.index', compact('projects'));
     }
@@ -29,6 +33,8 @@ class ProjectController extends Controller
      */
     public function show(Project $project): View
     {
+        Gate::authorize('view', $project);
+
         return view('pages.projects.show', compact('project'));
     }
 
@@ -39,7 +45,11 @@ class ProjectController extends Controller
      */
     public function create(): View
     {
-        return view('pages.projects.create');
+        Gate::authorize('create', Project::class);
+
+        $users = User::all();
+
+        return view('pages.projects.create', compact('users'));
     }
 
     /**
@@ -49,6 +59,8 @@ class ProjectController extends Controller
      */
     public function store(ProjectStoreRequest $request): RedirectResponse
     {
+        Gate::authorize('create', Project::class);
+
         $data = $request->validated();
         $data['is_active'] = $request->has('is_active') ? 1 : 0;
 
@@ -64,7 +76,11 @@ class ProjectController extends Controller
      */
     public function edit(Project $project): View
     {
-        return view('pages.projects.edit', compact('project'));
+        Gate::authorize('update', $project);
+
+        $users = User::all();
+
+        return view('pages.projects.edit', compact('project', 'users'));
     }
 
     /**
@@ -74,6 +90,8 @@ class ProjectController extends Controller
      */
     public function update(ProjectUpdateRequest $request, Project $project): RedirectResponse
     {
+        Gate::authorize('update', $project);
+
         $data = $request->validated();
         $data['is_active'] = $request->has('is_active') ? 1 : 0;
 
@@ -89,6 +107,8 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project): RedirectResponse
     {
+        Gate::authorize('delete', $project);
+
         $project->delete();
 
         return redirect()->route('projects.index', ['access' => 'yes']);
